Enable model access to admin users in Django admin

+++ api/serializers.py
@@ -31,8 +31,7 @@ class UserSerializer(serializers.ModelSerializer):
             validated_data["password"] = make_password(validated_data["password"])
 
         if validated_data.get("photo"):
-            if instance.photo:
-                instance.photo.delete()
+            if instance.photo: instance.photo.delete()
             validated_data["photo"].name = change_filename(validated_data["photo"])
 
         return super().update(instance, validated_data)
